front: extract helper function from ManchetteWithSpaceTimeChart
